Clarify and reorder koans related to Arrays and Generics
