+++ CSharp/AboutArrays.cs
@@ -67,41 +67,5 @@ namespace DotNetKoans.CSharp
 			Assert.Equal(new string[] { (string)FILL_ME_IN, (string)FILL_ME_IN }, array.Take(2).ToArray());
 			Assert.Equal(new string[] { (string)FILL_ME_IN, (string)FILL_ME_IN }, array.Skip(1).Take(2).ToArray());
         }
-
-        [Koan(5)]
-        public void PushingAndPopping()
-        {
-            var array = new[] { 1, 2 };
-            Stack stack = new Stack(array);
-            stack.Push("last");
-            Assert.Equal(FILL_ME_IN, stack.ToArray());
-            var poppedValue = stack.Pop();
-            Assert.Equal(FILL_ME_IN, poppedValue);
-            Assert.Equal(FILL_ME_IN, stack.ToArray());
-        }
-
-        [Koan(6)]
-        public void Shifting()
-        {
-            //Shift == Remove First Element
-            //Unshift == Insert Element at Beginning
-            //C# doesn't provide this natively. You have a couple
-            //of options, but we'll use the LinkedList<T> to implement
-            var array = new[] { "Hello", "World" };
-            var list = new LinkedList<string>(array);
-
-            list.AddFirst("Say");
-            Assert.Equal(FILL_ME_IN, list.ToArray());
-
-            list.RemoveLast();
-            Assert.Equal(FILL_ME_IN, list.ToArray());
-
-            list.RemoveFirst();
-            Assert.Equal(FILL_ME_IN, list.ToArray());
-
-            list.AddAfter(list.Find("Hello"), "World");
-            Assert.Equal(FILL_ME_IN, list.ToArray());
-        }
-
     }
 }
