+++ CSharp/CSharp.csproj
@@ -64,9 +64,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AboutArrayAssignment.cs" />
     <Compile Include="AboutArrays.cs" />
     <Compile Include="AboutAsserts.cs" />
+    <Compile Include="AboutAssignment.cs" />
     <Compile Include="AboutControlStatements.cs" />
     <Compile Include="AboutDelegates.cs" />
     <Compile Include="AboutGenericContainers.cs" />
